update for Jenkins 5
