+++ android/LOC_pytest.ini
@@ -1,0 +1,6 @@
+[pytest]
+markers =
+    uat: mark to test on uat.
+    int: mark to test on int.
+
+addopts = --driver Appium --variables caps.json -v -s --html=report.html --self-contained-html --disable-pytest-warnings
