--- android/temp_data.txt
@@ -1,2 +1,0 @@
-uat
-nemesis
