+++ android/test_uat.py
@@ -2,7 +2,7 @@ import pytest
 import os
 
 # LIST OF TESTS
-def atest_login_and_logout(login_model, debug_model, selenium):
+def test_login_and_logout(login_model, debug_model, selenium):
 	debug_model.switch_to_uat(selenium)
 	login_model.login_with_assert(selenium)
 	login_model.logout(selenium)
@@ -147,12 +147,12 @@ def atest_inbox_redirect_to_post_and_question_check(debug_model, login_model, in
 	login_model.login_only(selenium)
 	inbox_model.inbox_post_and_question_redirects(selenium)
 
-def test_dashboard_new_account_check(debug_model, login_model, dashboard_model, selenium):
+def atest_dashboard_new_account_check(debug_model, login_model, dashboard_model, selenium):
 	debug_model.switch_to_uat(selenium)
 	login_model.login_only_new_acc(selenium)
 	dashboard_model.new_acc_check(selenium)
 
-def test_dashboard_existing_account_check(debug_model, login_model, dashboard_model, selenium):
+def atest_dashboard_existing_account_check(debug_model, login_model, dashboard_model, selenium):
 	debug_model.switch_to_uat(selenium)
 	login_model.login_only(selenium)
 	dashboard_model.existing_acc_check(selenium)
